requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
async-timeout==5.0.1
attrs==25.3.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.0
contourpy==1.3.2
cryptography==44.0.3
cycler==0.12.1
datasets==3.1.0
dill==0.3.8
filelock==3.18.0
flwr==1.18.0
flwr-datasets==0.5.0
fonttools==4.58.0
frozenlist==1.6.0
fsspec==2024.9.0
grpcio==1.71.0
huggingface-hub==0.31.2
idna==3.10
iterators==0.0.2
Jinja2==3.1.6
kiwisolver==1.4.8
lightning-utilities==0.14.3
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
mdurl==0.1.2
mpmath==1.3.0
multidict==6.4.3
multiprocess==0.70.16
networkx==3.4.2
numpy==2.2.5
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
packaging==25.0
pandas==2.2.3
pathspec==0.12.1
pillow==11.2.1
propcache==0.3.1
protobuf==4.25.7
pyarrow==20.0.0
pycparser==2.22
pycryptodome==3.22.0
Pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
rich==13.9.4
seaborn==0.13.2
shellingham==1.5.4
six==1.17.0
sympy==1.14.0
tomli==2.2.1
tomli_w==1.2.0
torch==2.7.0
torchmetrics==1.7.1
torchvision==0.22.0
tqdm==4.67.1
triton==3.3.0
typer==0.12.5
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
xxhash==3.5.0
yarl==1.20.0
